tag: fix unified tagging and make unified2 compliant
Fixes issue 1854.

Logging of tagged packets in unified2 was not working as the
tag alert structure was never being initialized. Instead of
fixing that, making logging of tagged packets the same as in
Snort, where there is no event type for tagged packets, but
instead the event_id, and event_second field are used to
reference the event structure that initiated the tag.