flow: get rid of FLOW_TO_*_SEEN flags
As suggested by Victor Julien, we can now use todstpktcnt and
tosrcpktcnt directly instead of using these flags. This free 2
slots in the flags field of Flow structure.

This two flag were used to avoid to switch to established on an
ICMP error message. So by starting to increment the two values
only when we don't have a ICMP error message we mimic the same
behavior. On the logical side, it seems useless to increment a
flow for an error message.