detect-metadata: introduce target key
Suricata has currently no way to know which of the source and
destination IP address is the target of the attack. This is a
valuable information that is not available and thus can not be used
in the generated alerts.

The purpose of this patch is to introduce a new convention. The
target key will be used to define who from the source or destination
of the flow is the target of the attack.

The accepted syntax is:
 target [client|server]

This will be used later to provide the information in the alert
events.