dns: fix flood handling to avoid evasions
If a flow hit a DNS flood state with valid requests, new requests
that would match a block rule would be passed, or if not in IPS
mode, not logged/alerted.

To handle this the flood logic will not give up parsing anymore,
but instead the window is used to control how many outstanding
requests can be in memory, with a flood alert raised when number
of outstanding requests hits the limit, and the alert state cleared
when all DNS requests have been marked as replied (or reply lost).