app-layer: advance inspect_id while searching
The idea here is to advance the mimimum inspect_id as we come
across NULL transactions before hitting the first non-NULL
transaction.

This is a condition that can be seen in the worst case DNS
scenario where there is a flow of requests but no responses.

I'm not sure if this is the best place to do this, so its
still under review.