app-layer: don't disable on stream gaps
Allow app-layers to register that they can handle gaps.

Instead of sending a "signal" on a gap, send the gap flag on the next
segment, the app-layer "dispatcher" will check if the app-layer can
handle gaps or not, and abort if they can't, otherwise the data along
with a gap flag will be sent to the app-layer.

The idea is that the parser can then make the decision on what to do
with the gap.