Pcap Directory Mode (Feature #2222)
https://redmine.openinfosecfoundation.org/issues/2222

Pcap file mode that when passed a directory will process all files in that directory. If --pcap-file-continuous or continuous option is passed in json, the directory will be monitored  until the directory is moved/deleted, suricata is interrupted, or the pcap-interrupt command is used with unix command socket.

Engine state will not reset between files.