dhcp: don't allow ports to be configured
and remove some debug logging.

The DHCP ports configuration cannot be expressed in yaml
configuration file. For DHCP, the destination port determines
whether the request or response part should be called, not the
flow direction, as in most cases the flow will always be
to server. So for now, hard code the ports to 67 and 68.