app-layer: notify parsers of gaps if enabled
I parse can now set a flag that will tell the application
layer that it is capable of handling gaps. If enabled, and a
gap occurred, the app-layer needs to be prepared to accept
input that is NULL with a length, where the length is the
number of bytes lost. It is up to the app-layer to
determine if it can sync up with the input data again.