output-json-dns: summarize dns answer events
This permits to log a dns answer in one event rather than
a line for each answer.

To provide backward compatibility and avoid to break logging
systems it is necessary to enable eve-log.version option and set
it to 2.

Also it is possible to choice how a dns answer can be logged.
There are three method available: detailed, aggs, and both.

The following is a dns answer example when 'both' is set
as answer-type.

  "timestamp": "2017-07-31T15:01:17.885281+0200",
  "event_type": "dns",
  ...
  "dns": {
    "type": "answer",
    "id": 25394,
    "rcode": "NOERROR",
    "answers": [
      {
        "rrname": "notifications.google.com",
        "rrtype": "CNAME",
        "ttl": 3599,
        "rdata": "plus.l.google.com"
      },
      {
        "rrname": "plus.l.google.com",
        "rrtype": "A",
        "ttl": 299,
        "rdata": "216.58.205.174"
      }
    ],
    "aggs": {
      "A": [
        "216.58.205.174"
      ],
      "CNAME": [
        "plus.l.google.com"
      ]
    }
  }
}